fix: 🐛 factorize agent chat message accross channels
